QRcodeGenerator/src/requirements.txt
qrcodeqrcode[pil]==7.4.2
Pillow==10.0.0